feat(Active page indicator): Added Active page indicator, resolved suggested fix
Separated struct for more manageable code.
Renamed IndexPage -> DashboardPage for better readability.
