+++ inventorySystem/src/pages/DashboardPage.vue
