+++ inventorySystem/src/pages/InventoryPage.vue
@@ -1,14 +1,13 @@
 <template>
   <q-page class="row items-center justify-evenly">
     <h1>Inventory Page</h1>
-    <ProductCardGrid :products="inventoryProducts" />
+    <ProductCardGrid :products="productList" />
   </q-page>
 </template>
 <script setup lang="ts">
-import { ref } from 'vue'
-import ProductCardGrid, { type Product } from 'components/ProductCardGrid.vue'
-
-const inventoryProducts = ref<Product[]>([
+import ProductCardGrid from 'components/ProductCardGrid.vue'
+import type {Product} from '../types/products'
+const productList: Product[] = [
   {
     id: 1,
     name: 'Wireless Headphones',
@@ -42,21 +41,7 @@ const inventoryProducts = ref<Product[]>([
     image: 'https://picsum.photos/400/300?random=1'
   },
 
-
-])
+]
 </script>
 
-<style scoped>
-.sr-only {
-  position: absolute !important;
-  width: 1px !important;
-  height: 1px !important;
-  padding: 0 !important;
-  margin: -1px !important;
-  overflow: hidden !important;
-  clip: rect(0, 0, 0, 0) !important;
-  border: 0 !important;
-}
-
-</style>
