+++ inventorySystem/src/router/routes.ts
@@ -4,7 +4,7 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/',
     component: () => import('layouts/MainLayout.vue'),
-    children: [ { path: '', component: () => import('pages/IndexPage.vue')},
+    children: [ { path: '', component: () => import('pages/DashboardPage.vue')},
                 { path: 'inventory', component: () => import('pages/InventoryPage.vue') },
                 { path: 'sales', component: () => import('pages/SalesPage.vue') },
                 { path: 'contacts', component: () => import('pages/ContactsPage.vue') }
