+++ inventorySystem/src/types/menu-items.d.ts
@@ -2,4 +2,5 @@ export interface menuItemsProps {
   title: string;
   icon?: string;
   to: string;
+  exact?: boolean;
 }
