+++ inventorySystem/src/types/products.ts
@@ -1,0 +1,9 @@
+
+export interface Product {
+  id: number
+  name: string
+  description: string
+  price: number
+  stock: number
+  image: string
+}
